+++ Auto Neutron.py
@@ -5,9 +5,9 @@ import traceback
 
 from PyQt5 import QtWidgets, QtCore, QtGui
 
-import MainWindows
+import hub
 import popups
-from appinfo import *
+from appinfo import APP, ORG, APPID
 
 
 # https://stackoverflow.com/a/44352931
@@ -54,6 +54,6 @@ if __name__ == "__main__":
             os.mkdir(path[:path.rfind("/")])
         os.mkdir(path)
     settings = QtCore.QSettings(path + "/config.ini", QtCore.QSettings.IniFormat)
-    ui = MainWindows.Hub(settings)
+    ui = hub.Hub(settings)
     ui.startup()
     sys.exit(app.exec_())
